Added -c option to help resolve conflicts in git.
The new -c or --conflict option will rename the file with the conflict, <file>.c, to <file>.CONFLICT.c. It will
then restore your version of the file as <file>.c and the version being merged as <file>.THEIRS.c. It will then
run bbdiff on <file>.c and <file>.THEIRS.c so that you can resolve the merge changes in bbedit. There must be
conflict markers, (ie. <<<<<<<) in order for this option to be used. Otherwise git-bbedit-diff will complain
that no conflicts were found.

The new -c option is mutually exclusive to the -r or --revision option.